Manage command improved
