--- custom_command/management/commands/faker.py
@@ -1,8 +1,0 @@
-from django.core.management.base import BaseCommand
-
-
-class Command(BaseCommand):
-    help = 'The Zen of Python'
-
-    def handle(self, *args, **options):
-        import this
