+++ factory_boy/factory_model.py
@@ -1,29 +1,16 @@
-import random
-import os
-
+import factory.fuzzy
 import factory
-import django
 
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "DRF.settings")
-django.setup()
-
+from .factory_utils import (
+    generate_date_or_none,
+    clear_database,
+)
 
 from api.models import Person, Company, CompanyEmployee, Salary
 
 
-def clear_database(*args):
-    for model in args:
-        model.objects.all().delete()
-
-
-def generate_objects(count, obj, many=True, **kwargs):
-    if many:
-        return [obj(**kwargs) for _ in range(count)]
-    return obj(**kwargs)
-
-
 clear_database(Person, Company, CompanyEmployee, Salary)
+date_choice = generate_date_or_none()
 
 
 class PersonFactory(factory.django.DjangoModelFactory):
@@ -51,7 +38,13 @@ class CompanyEmployeeFactory(factory.django.DjangoModelFactory):
     supervisor = factory.SubFactory(PersonFactory)
     employee = factory.SubFactory(PersonFactory)
 
-    work_start_dt = factory.Faker('date')
+    work_start_dt = factory.Faker(
+        'date_between',
+        start_date="-5y",
+        end_date="today"
+    )
+
+    work_end_dt = factory.fuzzy.FuzzyChoice(date_choice)
 
 
 class SalaryFactory(factory.django.DjangoModelFactory):
@@ -62,23 +55,3 @@ class SalaryFactory(factory.django.DjangoModelFactory):
 
     salary = factory.Faker('pyint')
     date = factory.Faker('date_between', start_date="-5y", end_date="today")
-
-
-company_count = 5
-employees_count = 20
-
-companies = generate_objects(company_count, CompanyFactory)
-employees = generate_objects(employees_count, PersonFactory)
-
-for _ in range(employees_count):
-    companies_employees = generate_objects(
-        _,
-        CompanyEmployeeFactory,
-        company=random.choice(companies),
-        supervisor=random.choice(employees),
-        employee=random.choice(employees),
-        many=False
-    )
-
-
-# salary = SalaryFactory()
