--- test.py
@@ -1,6 +1,0 @@
-import random
-
-
-a = [1, 2, 8, 8, 9, 7, 2, 10, 3]
-
-print(random.choice(a))
